requirements.txt
flask==3.0.3
gunicorn==21.2.0
requests==2.32.3
cloudscraper==1.2.71
beautifulsoup4==4.12.3
lxml==5.3.0
rapidfuzz==3.9.7
discord.py==2.4.0
Pillow==10.4.0
imagehash>=4.3.1
